+++ app/services/cedar-loader.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-cli-cedar/services/cedar-loader';