+++ ember-cli-build.js
@@ -1,36 +1,27 @@
 'use strict';
 
 const EmberAddon = require('ember-cli/lib/broccoli/ember-addon');
+const publicPath = 'amcharts';
 
 module.exports = function(defaults) {
   let app = new EmberAddon(defaults, {
     // Add options here
     cedar: {
       amCharts: {
-        // publicPath: amCharts will be included at this path in the public folder
-        // use this if you are not loading amCharts from the CDN
-        publicPath: 'amcharts',
-        // imports: an array of amcharts files to be imported (bundled) in vendor files
-        imports: [
-          // amCharts base lib, this one is required
-          'amcharts.js',
-          // you will likely need at least one of the following:
-          'serial.js', // bar and line charts
-          'xy.js', // scatter charts
-          'pie.js', // pie and donut charts
-          'radar.js', // radar charts
-          // plugins are complely optioinal
-          // this one adds a download button on the chart
-          'plugins/export/export.js',
-          'plugins/export/export.css'
-        ]
+        // publicPath - amCharts will be included at this path in the public folder
+        // use this if you are NOT loading amCharts from the CDN
+        // i.e. you are going to either lazy-load amcharts from window.AmCharts_path
+        // or going to bundle amcharts into vendor.js
+        publicPath: publicPath,
+        // bundle - import amCharts files into vendor.js instead of lazy-loading them
+        // bundle: true
       }
     },
     // amCharts uses hardcoded paths to assets (like images and plugin files)
     // so you will need to make sure those don't get fingerprinted
     fingerprint: {
-      // you should  needs to be the same as publicPath above
-      exclude: ['amcharts']
+      // NOTE: this needs to be the same as publicPath above
+      exclude: [publicPath]
     }
   });